LayoutTests/fast/canvas/radialGradient-infinite-values-expected.txt

This test checks createRadialGradient with infinite values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.createRadialGradient(0, 0, 100, 0, 0, NaN) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, 0, 100, 0, 0, Infinity) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, 0, 100, 0, 0, -Infinity) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, 0, 100, 0, NaN, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, 0, 100, 0, Infinity, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, 0, 100, 0, -Infinity, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, 0, 100, NaN, 0, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, 0, 100, Infinity, 0, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, 0, 100, -Infinity, 0, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, 0, NaN, 0, 0, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, 0, Infinity, 0, 0, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, 0, -Infinity, 0, 0, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, NaN, 100, 0, 0, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, Infinity, 100, 0, 0, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(0, -Infinity, 100, 0, 0, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(NaN, 0, 100, 0, 0, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(Infinity, 0, 100, 0, 0, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS ctx.createRadialGradient(-Infinity, 0, 100, 0, 0, 100) threw exception Error: NotSupportedError: DOM Exception 9.
PASS successfullyParsed is true

TEST COMPLETE
